requirements.txt
# Django Web Framework
Django>=3.2,<4.0
djangorestframework>=3.12.0
django-cors-headers>=3.7.0
django-filter>=2.4.0

# Database
psycopg2-binary>=2.8.6  # PostgreSQL adapter
dj-database-url>=0.5.0  # Database URL parsing

# Authentication & Security
djangorestframework-simplejwt>=4.7.2
django-ratelimit>=3.0.0
cryptography>=3.4.8

# Email & Notifications
django-anymail>=8.4.0
celery>=5.2.0
redis>=3.5.3

# API Documentation
drf-yasg>=1.20.0
django-extensions>=3.1.5

# Development & Testing
pytest>=6.2.5
pytest-django>=4.4.0
pytest-cov>=2.12.1
factory-boy>=3.2.0
faker>=8.8.2

# Code Quality
black>=21.6b0
flake8>=3.9.2
isort>=5.9.3
mypy>=0.910
django-stubs>=1.9.0

# Monitoring & Logging
sentry-sdk>=1.3.1
django-health-check>=3.16.0

# Utilities
python-decouple>=3.4
Pillow>=8.3.2
python-dateutil>=2.8.2
pytz>=2021.1

# Production
gunicorn>=20.1.0
whitenoise>=5.3.0
django-storages>=1.11.1
boto3>=1.18.0  # For AWS S3 storage

# Environment
python-dotenv>=0.19.0
